GH-6 Code cleanup
