+++ handler.js
@@ -10,301 +10,26 @@
  * @author Gorilla Logic
  */
 
-const axios = require('axios');
-const { WebClient } = require('@slack/client'); // https://github.com/slackapi/node-slack-sdk
 const moment = require('moment');
-const { bambooAPIMethods } = require('./api/bamboo');
-const { parkingAPIMethods } = require('./api/parking');
-moment().format();
-require('moment-weekday-calc');
-
-// ================================ Lex Bot Helpers ===================================================================
-
-/**
- * These functions help you build responses that matches the output format expected by
- * Amazon Lex bots. Each structure is intended to be use on specific cases depending
- * on the conversation flow.
- * To know more about these output formats refer to Lex documentation:
- * {@link https://docs.aws.amazon.com/lex/latest/dg/lambda-input-response-format.html}.
- */
-
-function elicitSlot(
-  sessionAttributes,
-  intentName,
-  slots,
-  slotToElicit,
-  message
-) {
-  return {
-    sessionAttributes,
-    dialogAction: {
-      type: 'ElicitSlot',
-      intentName,
-      slots,
-      slotToElicit,
-      message,
-    },
-  };
-}
-
-function close(sessionAttributes, fulfillmentState, message) {
-  return {
-    sessionAttributes,
-    dialogAction: {
-      type: 'Close',
-      fulfillmentState,
-      message,
-    },
-  };
-}
-
-function delegate(sessionAttributes, slots) {
-  return {
-    sessionAttributes,
-    dialogAction: {
-      type: 'Delegate',
-      slots,
-    },
-  };
-}
-
-function buildValidationResult(isValid, violatedSlot, messageContent) {
-  if (messageContent == null) {
-    return {
-      isValid,
-      violatedSlot,
-    };
-  }
-  return {
-    isValid,
-    violatedSlot,
-    message: { contentType: 'PlainText', content: messageContent },
-  };
-}
-
-/**
- * Returns the given success message to the bot caller
- * @param {Object} intentRequest Intent request information
- * @param {function} callback Callback function to handle the response
- * @param {string} message The success message to return to the bot caller
- */
-const fulfillWithSuccess = (intentRequest, callback, message) => {
-  fulfillWithMessage(intentRequest, callback, message);
-};
-
-/**
- * Returns the given error message to the bot caller.
- * It returns a message with this pattern:
- *  - Sorry something failed. I got this error: [the error here].
- * @param {Object} intentRequest Intent request information
- * @param {function} callback Callback function to handle the response
- * @param {string} message The actual error to return to the bot caller
- */
-const fulfillWithError = (intentRequest, callback, message) => {
-  let error = `Sorry something failed. I got this error: ${message}.`;
-  fulfillWithMessage(intentRequest, callback, error);
-};
-
-const fulfillWithMessage = (intentRequest, callback, message) => {
-  // Fulfill the intent with a error response
-  callback(
-    close(intentRequest.sessionAttributes, 'Fulfilled', {
-      contentType: 'PlainText',
-      content: message,
-    })
-  );
-};
-
-// ================================ Helper Stuff ===================================================================
-/**
- * Collection of time-off request types from BambooHR.
- * The ids are specific to Gorilla Logic customized settings.
- * You can see the different options by looking at the UI and
- * using this API call:
- * GET {@link https://api.bamboohr.com/api/gateway.php/gorillalogic/v1/meta/time_off/types/}
- * With the proper API key.
- */
-const timeOffTypes = {
-  PTO: 1,
-  'Unpaid time off': 6,
-  'Bereavement Leave': 3,
-  'Paid Marriage Leave': 8,
-  'Travel Requests': 7,
-};
-
-/**
- * Collection of different responses to a user requesting a restaurant menu
- * from the different nearby restaurants that participate
- * on the Gorilla Logic lunch perk for Costa Rican gorillas.
- */
-const menus = {
-  nana: `The menu will be open (i.e. there will not be a specific menu for the Gorilla Logic Lunch Perk).
-Take a look at the menu here:
-https://res.cloudinary.com/greivinlopez/image/upload/v1565023471/menus/nana.jpg`,
-  mauros: `A specific menu will be offered for the Gorilla Logic Lunch Perk, which consultants are to request.
-Take a look at the menu here:
-https://res.cloudinary.com/greivinlopez/image/upload/v1565023470/menus/mauros.png`,
-  matsuri: `A specific menu will be offered for the Gorilla Logic Lunch Perk, which consultants are to request.
-Take a look at the menu here:
-https://res.cloudinary.com/greivinlopez/image/upload/v1565023471/menus/matsuri.png`,
-  dorado: `The dish of the day will be offered for the Gorilla Logic Lunch Perk.
-Take a look at the menu here:
-https://res.cloudinary.com/greivinlopez/image/upload/v1565023469/menus/dorado.png`,
-  chefilisimo: `This service may be used by following this link:
-https://docs.google.com/forms/d/1wD9W0fa1qa_p3AtXAqvf8l1STQ1HLjQ-4n1YIbaDrKo/edit?usp=sharing
-Orders are to be made the day before, and they will be delivered directly to the office.
-Meal orders will be delivered to the office at 11:30 a.m. on the 10th floor. Consultants will be
-responsible for picking up their orders.`,
-  cosi: `The -Taste Two- menu will be offered for the Gorilla Logic Lunch Perk.
-Take a look at the menu here:
-https://res.cloudinary.com/greivinlopez/image/upload/v1565794830/menus/cosi.jpg`,
-  cherry: `The menu will be open (i.e. there will not be a specific menu for the Gorilla Logic Lunch Perk).`,
-  bespoke: `The menu will be open (i.e. there will not be a specific menu for the Gorilla Logic Lunch Perk).`,
-  republica: `A specific menu will be offered for the Gorilla Logic Lunch Perk, which consultants are to request.
-Take a look at the menu here:
-https://res.cloudinary.com/greivinlopez/image/upload/v1565794833/menus/republica.jpg`,
-  paneevino: `A specific menu will be offered for the Gorilla Logic Lunch Perk, which consultants are to request. Sorry I don't have that menu yet.`,
-};
-
-const wellnessActivities = ['Yoga', 'ZenFit', 'MMA'];
-
-/**
- * Gets a human readable representation of the given date.
- * Using UTC -6 offset (Costa Rica timezone)
- * @param {Date} date A js Date instance
- * @return {string}
- */
-const humanDate = function(date) {
-  const COSTA_RICA_UTC_OFFSET = -6;
-  const m = moment(date).utcOffset(COSTA_RICA_UTC_OFFSET);
-  return m.calendar();
-};
-
-/**
- * Finds a person on the given employees object by matching the workEmail field..
- * @param {Object[]} employees Complete list of company employees.
- * @param {string} name The complete name of the employee to search for.
- * @return {Object} The object with the founded person's info. It returns undefined if not found.
- */
-const findPersonByEmail = function(employees, email) {
-  let findings = employees.filter(function(employee) {
-    return employee.workEmail === email;
-  });
-  if (findings.length > 0) {
-    return findings[0];
-  }
-};
-
-/**
- * Filter the given array of objects (timeOffItems) by using
- * the given PTO type. e.g. 'Vacations', 'PTO', 'Bereavement Leave', etc.
- * @param {Object[]} timeOffItems Array of JSON Objects with time off balances.
- * @param {string} ptoType A string describing the PTO type to use for filtering.
- * @return {object} The object with the given PTO time off balance. It returns undefined if not found.
- */
-const filterPTOs = function(timeOffItems, ptoType) {
-  let findings = timeOffItems.filter(function(timeOffItem) {
-    return timeOffItem.name === ptoType;
-  });
-  if (findings.length > 0) {
-    return findings[0];
-  }
-};
-
-/**
- * Randomly select a message from the given array
- * @param {array} messages An array of strings where to choose one.
- * @return {string}
- */
-const chooseOne = messages => {
-  return messages[Math.floor(Math.random() * messages.length)];
-};
-
-// ================================ Slack API ============================================================================
-// An access token (from your Slack app or custom integration - xoxp, xoxb, or xoxa)
-const slackApiToken = process.env.slackApiToken;
-
-/**
- * Extracts the user ID needed for calls to the Slack API as sent by the client
- * @param {Object} intentRequest Intent request information
- * @returns {string} The user ID for the Slack API coming on the client request
- */
-const getSlackUserId = intentRequest => {
-  /**
-   * @todo: Check if the request actually came from Slack.
-   * Under requestAttributes (from event) you can check the presence of
-   * x-amz-lex:channel-type.
-   * The value will be Slack if the user comes from slack.
-   */
-  let userId = intentRequest.userId;
-  userId = userId.slice(userId.lastIndexOf(':') + 1);
-  return userId;
-};
-
-/**
- * Gets a promise to request the user information from Slack API
- * associated to the given user ID.
- * @param {string} userId Slack user ID
- */
-const getSlackUserInfo = userId => {
-  const web = new WebClient(slackApiToken);
-  return web.users.info({ user: userId });
-};
-
-/**
- * Retrieve the profile information from Slack for the given ID
- * @param {string} userId Slack user ID
- */
-const getSlackUserProfile = userId => {
-  const web = new WebClient(slackApiToken);
-  return web.users.profile.get({ user: userId });
-};
-
-// ================================ Numbers API ===============================
-const numbersAPI = axios.create({
-  baseURL: 'http://numbersapi.com/',
-  timeout: 6000,
-  headers: {
-    accept: 'text/plain',
-    'accept-encoding': 'gzip, deflate',
-    'accept-language': 'en-US,en;q=0.8',
-  },
-});
-
-// ================================ Google Calendar ===============================
 const { google } = require('googleapis');
 
-const calendarId = '6i7h19gftsao0fl18nibeukts8@group.calendar.google.com';
-const serviceAccount = {
-  type: 'service_account',
-  project_id: 'winston-2d06e',
-  private_key_id: 'c34bced7e2e96c08b067de38f1fd418eee4c5aa6',
-  private_key:
-    '-----BEGIN PRIVATE KEY-----\nMIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQDLoqVrN4RRB6y/\nd5mrovpsI7Kcx4SKN6gwZ0ohcBTim+p7f5Qz3e09/UcMY0fecNZy9zGzpadj3FkW\nYkZRufEdDZ8JbxaT1TYepioQMQecXC1B53hXURfzhbBlR9iaYmz6A2Cnu7jC24zF\ncNOKBFqhUgECxokJASwZ32S+V2eB32B+Mz6Mk6gx0/9ASOKtmlV6myPQlIOn1oug\nx9Y+m15QDL0wf0CwgR957vRkZOGibdAqVzCHoKOsKO1i5A8ktehrK6Htp4yT/sCT\nlk4Ocw+mP2TJ5d2mf3C8sXfRZ0lOP9HGvhTuY4ohjB7J4uIiDo2CDN71RhoGVv6I\n/Ii1iIYLAgMBAAECggEAF5u9rdZ74qEPKAuHp1x3x/3DgWWoOhYFy8bN016ZLUSC\nRNsfpXu/O4aXrBn+hQ1z71Q1UOyNoYg0U+IE4Qi7qXCg1Ykphu4bHiaJjeMBTOFq\n5c8eJpFiQ3tmXfecZxmuJLYriIa8/upRQbq467FhiH5nW5kE3iRipoU9usuLVoUI\nDMsEtZLey5xHteYdgRMu3dmSSofvFs6cPWYlcxkl6sY6FGY4aLlxGuURXVHwz7OV\nnVN1DY7TmqnzyaPPiBY9X9Of24Ieoalx8He5hCE3MyVp+pNbD8W7IrVAUQ8OI8WT\nB4O2FnEGtqDbLLElr/MJc+ZSVHPQlqZ5Ec0csWPPUQKBgQD7bmoWa585BC77oY1o\nu+Bt5no23yEezPLt3WGCradFMhXeGuaxX7KyaAQoyrLc5ysvsskv/tScvcmC9bIQ\nsARD5kIRISsExJaoRrOzgrkFLNQuPrwTY0w6EDDkSUy7Cr6zLl4avhrIj/6luLf0\n/WHZcK85nU80Gz5eN+GCbpzsNQKBgQDPVeYVCdsEbCSIGF7F6mziWucrvl1N2fSy\nc1T6xtegf5oeioA0yMI8IuSFCLQroBceUEzCcADAU98HpvVxFUhnVJAWYymkDWY4\nt0yaR/qek63Ar4G9wg+85GtzLQTRJnh3k10GlR+t1ggOErTrn7KEzG/3mJU4c2Hz\nqgyWVOtxPwKBgFK+w6sdQAlDlzK4uBD/w5xnfVERnZTFXi1p7e1TuPjMwuWVdNZq\n6CROlw/VpYeNjGjmLc/dzZrH2sl+kltrkfvGi3eXstCe4fX/TqSFIN7TIgEWJE89\nHpj9OxnmBZqIrrnJSK8EYUwgF23ynJfDy7ADpacQXkCHtr7LJSwHux0BAoGANVoF\nOqSulxS80chzmroYCqcZlUgG+mF7de/3jhBcrBApuHIrS2ndkVvpMdJrik5HpKVF\nG0N27gam/XEY92BioTRBX4Fk1bIc+7svGwQqmWOm8k43NH0bT5/3hkuvUaJ0nVBu\nP6Nf43uP1w85Puh5J3Lz3OKGg2AtdGXtl0GC1EECgYAPdtuhlLhCS9JW9f/IrvGo\n24s2nNxHzXywyTC5rfGApFxHngTNrVzYsLEFgMEvF5I8sQTd9pa7FBAdIS+rTvLU\nOZT4S+67n7Q8fHbMayiM3qVmNEw86iq8d4gj8npGAn1pYP9u4lsK0byoaxe5kki7\nCsc4Eg52IgiBEiJY7TTtFg==\n-----END PRIVATE KEY-----\n',
-  client_email: 'wellness-activities@winston-2d06e.iam.gserviceaccount.com',
-  client_id: '103934030182041620924',
-  auth_uri: 'https://accounts.google.com/o/oauth2/auth',
-  token_uri: 'https://oauth2.googleapis.com/token',
-  auth_provider_x509_cert_url: 'https://www.googleapis.com/oauth2/v1/certs',
-  client_x509_cert_url:
-    'https://www.googleapis.com/robot/v1/metadata/x509/wellness-activities%40winston-2d06e.iam.gserviceaccount.com',
-};
+const { bambooAPIMethods } = require('./api/bamboo');
+const { parkingAPIMethods } = require('./api/parking');
+const { slackAPIMethods } = require('./api/slack');
+const { numbersAPIMethods } = require('./api/numbers');
+const googleCalendar = require('./api/google/calendar');
 
-const serviceAccountAuth = new google.auth.JWT({
-  email: serviceAccount.client_email,
-  key: serviceAccount.private_key,
-  scopes: 'https://www.googleapis.com/auth/calendar',
-});
+const lex = require('./helpers/lex');
+const bamboo = require('./helpers/bamboo');
+const slack = require('./helpers/slack');
+const foodBonus = require('./helpers/foodBonus');
+const wellness = require('./helpers/wellness');
 
-const getUpcomingWellnessEvent = function(activity, events) {
-  const upcoming = events.filter(event => {
-    return event.summary === activity;
-  });
-  if (upcoming.length > 0) {
-    return upcoming[0];
-  }
-  return null;
-};
+const dateFormatters = require('./helpers/formatters/date');
+const random = require('./helpers/random');
+
+moment().format();
+require('moment-weekday-calc');
 
 // ================================ Functions to handle intents =============================================================
 
@@ -314,32 +39,33 @@ const getUpcomingWellnessEvent = function(activity, events) {
  * @param {function} callback Callback function to handle the response
  */
 const getTimeOffBalance = async function(intentRequest, callback) {
-  const userId = getSlackUserId(intentRequest);
+  const userId = slack.getSlackUserId(intentRequest);
   // Request Slack API to extract the email from the user Id
   let userProfile = null;
   let person = null;
   try {
-    userProfile = await getSlackUserProfile(userId);
+    userProfile = await slackAPIMethods.getSlackUserProfile(userId);
     const getEmployeesResult = await bambooAPIMethods.getEmployees();
 
-    let slackEmail = userProfile.profile.email;
-    let employees = getEmployeesResult.data.employees;
-    person = findPersonByEmail(employees, slackEmail);
+    const slackEmail = userProfile.profile.email;
+    const { employees } = getEmployeesResult.data;
+    const [personResult] = employees.filter(e => e.workEmail === slackEmail);
 
-    if (person === undefined) {
-      let err = `Sorry, ${slackEmail} could not be found`;
+    if (personResult === undefined) {
+      const err = `Sorry, ${slackEmail} could not be found`;
       throw err;
     }
+    person = personResult;
   } catch (error) {
     console.log(error);
-    fulfillWithError(intentRequest, callback, error);
+    lex.fulfillWithError(intentRequest, callback, error);
     return;
   }
 
-  let today = new Date();
-  let dd = today.getDate();
-  let mm = today.getMonth() + 1; //January is 0!
-  let yyyy = today.getFullYear();
+  const today = new Date();
+  const dd = today.getDate();
+  const mm = today.getMonth() + 1; //January is 0!
+  const yyyy = today.getFullYear();
 
   try {
     const response = await bambooAPIMethods.calculateTimeOffBalance(
@@ -348,8 +74,8 @@ const getTimeOffBalance = async function(intentRequest, callback) {
       mm,
       dd
     );
-    let ptos = filterPTOs(response.data, 'PTO'); //Filter time off by PTO
-    let employeeName = person.displayName;
+    const [ptos] = response.data.filter(t => t.name === 'PTO');
+    const employeeName = person.displayName;
     if (ptos === undefined) {
       console.log(`Sorry, could not found PTO information for ${employeeName}`);
       return;
@@ -357,10 +83,10 @@ const getTimeOffBalance = async function(intentRequest, callback) {
     // Success - answer to the user with the PTO days
     const message = `${employeeName} you have ${ptos.balance} ${ptos.units} left`;
     console.log(message);
-    fulfillWithSuccess(intentRequest, callback, message);
+    lex.fulfillWithSuccess(intentRequest, callback, message);
   } catch (error) {
     console.log(error);
-    fulfillWithError(intentRequest, callback, error);
+    lex.fulfillWithError(intentRequest, callback, error);
   }
 };
 
@@ -371,8 +97,8 @@ const getTimeOffBalance = async function(intentRequest, callback) {
  */
 const tellAJokeAboutChuckNorris = (intentRequest, callback) => {
   const jokes = require('./chucknorris.json');
-  const joke = chooseOne(jokes);
-  fulfillWithSuccess(intentRequest, callback, joke);
+  const joke = random.shuffleAndPickOne(jokes);
+  lex.fulfillWithSuccess(intentRequest, callback, joke);
 };
 
 /**
@@ -386,9 +112,9 @@ const tellAJokeAboutChuckNorris = (intentRequest, callback) => {
 const validateTimeOffRequest = function(typeOfTimeOff, startDate, endDate) {
   if (
     typeOfTimeOff &&
-    Object.keys(timeOffTypes).indexOf(typeOfTimeOff) === -1
+    Object.keys(bamboo.timeOffTypes).indexOf(typeOfTimeOff) === -1
   ) {
-    return buildValidationResult(
+    return lex.buildValidationResult(
       false,
       'typeOfTimeOff',
       `Sorry I can't create a time off request for ${typeOfTimeOff}, would you like a different type of time off?`
@@ -396,7 +122,7 @@ const validateTimeOffRequest = function(typeOfTimeOff, startDate, endDate) {
   }
   if (startDate) {
     if (new Date(startDate) < new Date()) {
-      return buildValidationResult(
+      return lex.buildValidationResult(
         false,
         'startDate',
         `I can't schedule a time off request in the past! Can you provide a different date?`
@@ -405,14 +131,14 @@ const validateTimeOffRequest = function(typeOfTimeOff, startDate, endDate) {
   }
   if (endDate) {
     if (new Date(endDate) < new Date(startDate)) {
-      return buildValidationResult(
+      return lex.buildValidationResult(
         false,
         'endDate',
         'Your return date must be after your leave date! Can you try a different date?'
       );
     }
   }
-  return buildValidationResult(true, null, null);
+  return lex.buildValidationResult(true, null, null);
 };
 
 /**
@@ -422,16 +148,18 @@ const validateTimeOffRequest = function(typeOfTimeOff, startDate, endDate) {
  * @param {function} callback Callback function to handle the response
  */
 const createTimeOffRequest = async function(intentRequest, callback) {
-  const typeOfTimeOff = intentRequest.currentIntent.slots.typeOfTimeOff;
-  const startDate = intentRequest.currentIntent.slots.startDate;
-  const endDate = intentRequest.currentIntent.slots.endDate;
+  const {
+    typeOfTimeOff,
+    startDate,
+    endDate,
+  } = intentRequest.currentIntent.slots;
   // Session attributes from the intent.
   var outputSessionAttributes = intentRequest.sessionAttributes;
 
   if (intentRequest.invocationSource === 'DialogCodeHook') {
     // Perform basic validation on the supplied input slots.
     // Use the elicitSlot dialog action to re-prompt for the first violation detected.
-    const slots = intentRequest.currentIntent.slots;
+    const { slots } = intentRequest.currentIntent;
     const validationResult = validateTimeOffRequest(
       typeOfTimeOff,
       startDate,
@@ -440,7 +168,7 @@ const createTimeOffRequest = async function(intentRequest, callback) {
     if (!validationResult.isValid) {
       slots[`${validationResult.violatedSlot}`] = null;
       callback(
-        elicitSlot(
+        lex.elicitSlot(
           intentRequest.sessionAttributes,
           intentRequest.currentIntent.name,
           slots,
@@ -456,8 +184,8 @@ const createTimeOffRequest = async function(intentRequest, callback) {
       if (!outputSessionAttributes) {
         outputSessionAttributes = { confirmationPrompt: '' };
       }
-      let displayStart = moment(startDate).format('dddd, MMMM Do YYYY');
-      let displayEnd = moment(endDate).format('dddd, MMMM Do YYYY');
+      const displayStart = moment(startDate).format('dddd, MMMM Do YYYY');
+      const displayEnd = moment(endDate).format('dddd, MMMM Do YYYY');
 
       let confirmationPrompt = `Can you confirm your ${typeOfTimeOff} request from ${displayStart} to ${displayEnd}?`;
       if (startDate === endDate) {
@@ -467,30 +195,29 @@ const createTimeOffRequest = async function(intentRequest, callback) {
       outputSessionAttributes.confirmationPrompt = confirmationPrompt;
     }
     callback(
-      delegate(outputSessionAttributes, intentRequest.currentIntent.slots)
+      lex.delegate(outputSessionAttributes, intentRequest.currentIntent.slots)
     );
     return;
   }
 
-  const userId = getSlackUserId(intentRequest);
+  const userId = slack.getSlackUserId(intentRequest);
   // Request Slack API to extract the email from the user Id
   let userProfile = null;
   let person = null;
   try {
-    userProfile = await getSlackUserProfile(userId);
+    userProfile = await slackAPIMethods.getSlackUserProfile(userId);
     const getEmployeesResult = await bambooAPIMethods.getEmployees();
 
     const slackEmail = userProfile.profile.email;
-    const employees = getEmployeesResult.data.employees;
-    person = findPersonByEmail(employees, slackEmail);
-
+    const { employees } = getEmployeesResult.data;
+    [person] = employees.filter(e => e.workEmail === slackEmail);
     if (person === undefined) {
       const err = `Sorry, ${slackEmail} could not be found`;
       throw err;
     }
   } catch (error) {
     console.log(error);
-    fulfillWithError(intentRequest, callback, error);
+    lex.fulfillWithError(intentRequest, callback, error);
     return;
   }
 
@@ -501,7 +228,7 @@ const createTimeOffRequest = async function(intentRequest, callback) {
     // Filter non-holidays elements from this
     // response and put the dates for company
     // holidays into an array.
-    const items = response.filter(function(item) {
+    const items = response.filter(item => {
       return item.type === 'holiday';
     });
     const holidays = [];
@@ -519,7 +246,7 @@ const createTimeOffRequest = async function(intentRequest, callback) {
       exclusions: holidays,
     });
 
-    const timeOffTypeValue = timeOffTypes[typeOfTimeOff];
+    const timeOffTypeValue = bamboo.timeOffTypes[typeOfTimeOff];
 
     try {
       const response = await bambooAPIMethods.sendTimeOffRequest(
@@ -535,10 +262,10 @@ const createTimeOffRequest = async function(intentRequest, callback) {
         content = `OK ${person.firstName}, I have sent your request. Please wait for approval from ${approver}.`;
       }
       // Fulfill the request
-      fulfillWithSuccess(intentRequest, callback, content);
+      lex.fulfillWithSuccess(intentRequest, callback, content);
     } catch (error) {
       console.log(error);
-      fulfillWithError(intentRequest, callback, error);
+      lex.fulfillWithError(intentRequest, callback, error);
     }
   }
 };
@@ -552,23 +279,22 @@ const createTimeOffRequest = async function(intentRequest, callback) {
  * @param {function} callback Callback function to handle the response
  */
 const updateLicensePlate = async function(intentRequest, callback) {
-  let previousPlate = intentRequest.currentIntent.slots.previousPlate;
-  let newPlate = intentRequest.currentIntent.slots.newPlate;
+  const { previousPlate, newPlate } = intentRequest.currentIntent.slots;
 
   if (!previousPlate || !newPlate) {
-    let errorMessage = `Something failed with arguments previousPlate: ${previousPlate} or newPlate: ${newPlate}`;
+    const errorMessage = `Something failed with arguments previousPlate: ${previousPlate} or newPlate: ${newPlate}`;
     console.log(errorMessage);
-    fulfillWithError(intentRequest, callback, errorMessage);
+    lex.fulfillWithError(intentRequest, callback, errorMessage);
   }
 
-  const userId = getSlackUserId(intentRequest);
+  const userId = slack.getSlackUserId(intentRequest);
   let userInfo = null;
   try {
     // Retrieve user information from Slack using the SDK
-    userInfo = await getSlackUserInfo(userId);
+    userInfo = await slack.getSlackUserInfo(userId);
   } catch (error) {
     console.log(error);
-    fulfillWithError(
+    lex.fulfillWithError(
       intentRequest,
       callback,
       "Failed to retrieve user's info from Slack"
@@ -584,10 +310,10 @@ const updateLicensePlate = async function(intentRequest, callback) {
     });
     console.log(res.data);
     const message = `Your plate number was updated to ${res.data.newPlate}.`;
-    fulfillWithSuccess(intentRequest, callback, message);
+    lex.fulfillWithSuccess(intentRequest, callback, message);
   } catch (error) {
     console.log(error);
-    fulfillWithError(intentRequest, callback, error.response.data.error);
+    lex.fulfillWithError(intentRequest, callback, error.response.data.error);
   }
 };
 
@@ -601,22 +327,21 @@ const countGorillas = async function(intentRequest, callback) {
   try {
     const getEmployeesResult = await bambooAPIMethods.getEmployees();
     const employeesCount = getEmployeesResult.data.employees.length + 1;
-
-    numbersAPI
-      .get(`/${employeesCount}/trivia?notfound=floor&fragment`)
-      .then(res => {
-        console.log(res.data);
-        const message = `We are ${employeesCount} souls which is close to ${res.data}.`;
-        fulfillWithSuccess(intentRequest, callback, message);
-      })
-      .catch(error => {
-        console.log(error.data);
-        const message = `We are ${employeesCount} souls.`;
-        fulfillWithSuccess(intentRequest, callback, message);
-      });
+    try {
+      const res = await numbersAPIMethods.getTriviaForEmployeeCount(
+        employeesCount
+      );
+      console.log(res.data);
+      const message = `We are ${employeesCount} souls which is close to ${res.data}.`;
+      lex.fulfillWithSuccess(intentRequest, callback, message);
+    } catch (error) {
+      console.log(error.data);
+      const message = `We are ${employeesCount} souls.`;
+      lex.fulfillWithSuccess(intentRequest, callback, message);
+    }
   } catch (error) {
     console.log(error);
-    fulfillWithError(intentRequest, callback, error);
+    lex.fulfillWithError(intentRequest, callback, error);
   }
 };
 
@@ -626,16 +351,16 @@ const countGorillas = async function(intentRequest, callback) {
  * @param {function} callback Callback function to handle the response
  */
 const getRestaurantMenu = function(intentRequest, callback) {
-  const restaurant = intentRequest.currentIntent.slots.restaurant;
+  const { restaurant } = intentRequest.currentIntent.slots;
 
   // eslint-disable-next-line no-prototype-builtins
-  if (!restaurant || !menus.hasOwnProperty(restaurant)) {
+  if (!restaurant || !foodBonus.menus.hasOwnProperty(restaurant)) {
     const errorMessage = `Sorry, I'm not aware of the menu for ${restaurant}`;
-    fulfillWithSuccess(intentRequest, callback, errorMessage);
+    lex.fulfillWithSuccess(intentRequest, callback, errorMessage);
   }
 
-  const response = menus[restaurant];
-  fulfillWithSuccess(intentRequest, callback, response);
+  const response = foodBonus.menus[restaurant];
+  lex.fulfillWithSuccess(intentRequest, callback, response);
 };
 
 const getInformationAboutWellnessActivities = async function(
@@ -643,30 +368,31 @@ const getInformationAboutWellnessActivities = async function(
   callback
 ) {
   const activity = intentRequest.currentIntent.slots.wellness;
-  if (!activity || wellnessActivities.indexOf(activity) === -1) {
+  if (!activity || wellness.wellnessActivities.indexOf(activity) === -1) {
     const errorMessage = `Sorry I didn't get the intended activity name`;
-    fulfillWithSuccess(intentRequest, callback, errorMessage);
+    lex.fulfillWithSuccess(intentRequest, callback, errorMessage);
     return;
   }
   const calendar = google.calendar('v3');
   const timeMin = moment(new Date());
   try {
     const response = await calendar.events.list({
-      auth: serviceAccountAuth,
-      calendarId,
+      auth: googleCalendar.serviceAccountAuth,
+      calendarId: googleCalendar.calendarId,
       timeMin: timeMin.toISOString(),
       maxResults: 10,
       singleEvents: true,
       orderBy: 'startTime',
     });
     const events = response.data.items;
-    const event = getUpcomingWellnessEvent(activity, events);
+    const [event] = events.filter(e => e.summary === activity);
     if (!event) {
       const noEventsFoundMessage = `No events found`;
-      fulfillWithSuccess(intentRequest, callback, noEventsFoundMessage);
+      lex.fulfillWithSuccess(intentRequest, callback, noEventsFoundMessage);
       return;
     }
-    const readableActivity = `Next ${activity} activity will be ${humanDate(
+    const readableActivity = `Next ${activity} activity will be ${dateFormatters.humanDate(
+      moment,
       event.start.dateTime
     )}`;
     const additionalText = event.description
@@ -674,10 +400,10 @@ const getInformationAboutWellnessActivities = async function(
       : `To stay up to date with the coming activities visit the company's portal.`;
     const eventsUrl = 'https://band.gorillalogic.com/events/';
     const message = `${readableActivity}\n${additionalText}\nSee events:\n${eventsUrl}`;
-    fulfillWithSuccess(intentRequest, callback, message);
+    lex.fulfillWithSuccess(intentRequest, callback, message);
   } catch (error) {
     console.log(error);
-    fulfillWithError(intentRequest, callback, error);
+    lex.fulfillWithError(intentRequest, callback, error);
   }
 };
 
@@ -725,7 +451,6 @@ function dispatch(intentRequest, callback) {
 module.exports.winstonbot = (event, context, callback) => {
   try {
     console.log(`event.bot.name=${event.bot.name}`);
-
     /**
      * Sanity check to prevent others for invoking this Lambda function
      * outside the intended bot
@@ -733,7 +458,6 @@ module.exports.winstonbot = (event, context, callback) => {
     if (event.bot.name !== 'Winston') {
       callback('Invalid Bot Name');
     }
-
     dispatch(event, response => callback(null, response));
   } catch (err) {
     callback(err);
