+++ helpers/foodBonus/index.js
@@ -1,0 +1,35 @@
+/**
+ * Collection of different responses to a user requesting a restaurant menu
+ * from the different nearby restaurants that participate
+ * on the Gorilla Logic lunch perk for Costa Rican gorillas.
+ */
+const menus = {
+  nana: `The menu will be open (i.e. there will not be a specific menu for the Gorilla Logic Lunch Perk).
+Take a look at the menu here:
+https://res.cloudinary.com/greivinlopez/image/upload/v1565023471/menus/nana.jpg`,
+  mauros: `A specific menu will be offered for the Gorilla Logic Lunch Perk, which consultants are to request.
+Take a look at the menu here:
+https://res.cloudinary.com/greivinlopez/image/upload/v1565023470/menus/mauros.png`,
+  matsuri: `A specific menu will be offered for the Gorilla Logic Lunch Perk, which consultants are to request.
+Take a look at the menu here:
+https://res.cloudinary.com/greivinlopez/image/upload/v1565023471/menus/matsuri.png`,
+  dorado: `The dish of the day will be offered for the Gorilla Logic Lunch Perk.
+Take a look at the menu here:
+https://res.cloudinary.com/greivinlopez/image/upload/v1565023469/menus/dorado.png`,
+  chefilisimo: `This service may be used by following this link:
+https://docs.google.com/forms/d/1wD9W0fa1qa_p3AtXAqvf8l1STQ1HLjQ-4n1YIbaDrKo/edit?usp=sharing
+Orders are to be made the day before, and they will be delivered directly to the office.
+Meal orders will be delivered to the office at 11:30 a.m. on the 10th floor. Consultants will be
+responsible for picking up their orders.`,
+  cosi: `The -Taste Two- menu will be offered for the Gorilla Logic Lunch Perk.
+Take a look at the menu here:
+https://res.cloudinary.com/greivinlopez/image/upload/v1565794830/menus/cosi.jpg`,
+  cherry: `The menu will be open (i.e. there will not be a specific menu for the Gorilla Logic Lunch Perk).`,
+  bespoke: `The menu will be open (i.e. there will not be a specific menu for the Gorilla Logic Lunch Perk).`,
+  republica: `A specific menu will be offered for the Gorilla Logic Lunch Perk, which consultants are to request.
+Take a look at the menu here:
+https://res.cloudinary.com/greivinlopez/image/upload/v1565794833/menus/republica.jpg`,
+  paneevino: `A specific menu will be offered for the Gorilla Logic Lunch Perk, which consultants are to request. Sorry I don't have that menu yet.`,
+};
+
+module.exports = { menus };
