+++ helpers/wellness/index.js
@@ -1,0 +1,3 @@
+const wellnessActivities = ['Yoga', 'ZenFit', 'MMA'];
+
+module.exports = { wellnessActivities };
